SAF-69: Static: Make sidebar responsive
When the screen width is small or narrower, the sidebar disappears and
only reappears via a click of a hamburger menu button at the top left
corner of the page. On click of this button, the sidebar slides out from
the left side.

Modified the Home component to horizontally center the title of the
component, so that the user can see what page they are on, even if it is
just a placeholder.