SAF-67: Static: Route sidebar buttons
Routed the sidebar buttons to their corresponding pages. As a result,
the pages were created but only contain placeholder elements (i.e., the
title of the page).

Additionally, simplified the Home page to match with the other pages.
That is, the dummy button was removed and the title was horizontally
centered.