SAF-111: Static: Move pages to its own component
Moved all pages to its own component. By doing so, the useLocation hook
is available in the new Routes component. With this hook, the Pages
component is able to conditionally render either the Login page (i.e., a
page without a sidebar) or any other page based on the URL.