Fix margin of Home page
In a previous commit, the padding of inner content was removed on small
screen sizes. This is so that the contents of each page can fully use
the entire screen size. However, the home page has content that should
not be directly touching the edges of the screen (i.e., text content).
Therefore, added margin just for the Home page.