Add filter box to mobile view
Added the filter box to the mobile view. The filter box is the same one
used in the desktop view. Additionally, modified the filter box to be in
a modal. By default, it is hidden, and it is only displayed on click of
a filter button on the bottom left of the display.